refactor: move withExpiringActivity to WireSystem
