+++ wire-ios-request-strategy/Sources/Synchronization/Decoding/EventDecoder.swift
@@ -152,59 +152,6 @@ extension EventDecoder {
         )
     }
 
-    func withExpiringActivity(reason: String, block: @escaping () async throws -> Void) async throws {
-        let manager = ExpiringActivityManager()
-        try await manager.withExpiringActivity(reason: reason, block: block)
-    }
-
-    actor ExpiringActivityManager {
-        var task: Task<Void, Error>?
-
-        func withExpiringActivity(reason: String, block: @escaping () async throws -> Void) async throws {
-            try await withCheckedThrowingContinuation { continuation in
-                ProcessInfo.processInfo.performExpiringActivity(withReason: reason) { expiring in
-                    if !expiring {
-                        let semaphore = DispatchSemaphore(value: 0)
-                        Task {
-                            do {
-                                try await self.startWork(block: block, semaphore: semaphore).value
-                                WireLogger.backgroundActivity.debug("Expiring activity completed: \(reason)")
-                                continuation.resume()
-                            } catch {
-                                WireLogger.backgroundActivity.debug("Expiring activity ended with an error: \(error)")
-                                continuation.resume(throwing: error)
-                            }
-
-                        }
-                        semaphore.wait()
-                    } else {
-                        WireLogger.backgroundActivity.warn("Background activity is expiring: \(reason)")
-                        Task {
-                            await self.stopWork()
-                        }
-                    }
-                }
-            }
-        }
-
-        func startWork(block: @escaping () async throws -> Void, semaphore: DispatchSemaphore) -> Task<Void, Error> {
-            let task = Task {
-                defer {
-                    WireLogger.backgroundActivity.debug("Releasing semaphore")
-                    semaphore.signal()
-                }
-                try await block()
-            }
-            self.task = task
-            return task
-        }
-
-        func stopWork() {
-            self.task?.cancel()
-            self.task = nil
-        }
-    }
-
     /// Decrypts and stores the decrypted events as `StoreUpdateEvent` in the event database.
     /// The encryption context is only closed after the events have been stored, which ensures
     /// they can be decrypted again in case of a crash.
